branches/main devicesTutorial example1: added ICE networking example with SensableHD
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@679 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ examples/devicesTutorial/example1/main.cpp
@@ -25,7 +25,7 @@ int main(void)
     cmnClassRegister::SetLoD("devSensableHD", CMN_LOG_LOD_VERY_VERBOSE);
 
     // create our two tasks
-    const long PeriodDisplay = 10; // in milliseconds
+    const double PeriodDisplay = 10.0; // in milliseconds
     mtsTaskManager * taskManager = mtsTaskManager::GetInstance();
     displayTask * displayTaskObject =
         new displayTask("DISP", PeriodDisplay * cmn_ms);
